#23 refactor: 기존 CloneElement를 통해 구현한 구독 취소 버튼을 훅으로 수정
